fixed comment checking
the way i was doing it previously was a really hacky way which essentially just checked if "anonbot.wl" was mentioned anywhere on the given URL and only posted if it wasn't. but with this method, a comment would still be posted if someone mentioned "anonbot.wl" anywhere in their post. so i dug through the instagram API and found a way to get who posted a post from a given link. now it truly will only comment on anonbot posts