added ipbanning functionality
I figured I'd wait until someone actually breaks enough rules to be IPbanned to add this. Today, someone spammed a bunch and there was no way to stop them other than to fight with them and keep deleting thier posts while trying to implement this functionality. This was just thrown together to ban that one person. working on adding a blacklist and hiding it from the public so as not to expose any IP addresses