completely redesigned log system 🎉
Instead of writing to a json file, it now updates an Airtable field. This now makes it accessible to all moderators who are invited to see it (keeping IPs closed off from the public since it's no longer just in a json file) and Airtable's API is so beautiful and simple it literally took me 10 minutes. I'm in love.